Refactor breaking up app struct
